% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reganalysis.R
\name{reg_analysis}
\alias{reg_analysis}
\title{reg_analysis function}
\usage{
reg_analysis()
}
\value{
Linear regression model
}
\description{
Data Analisys: linear regression model to evaluate the "tranquilizer use" variation due to the "alcohol use" change.

It's necessary to import other packages if the are not installed.

The output shows some significant estimates which allow us to understand how the model significative is.
}
\author{
Francesca Grazia Radatti \email{francescaradatti97@gmail.com}
}
